Segregate generic physical handling and PX actor code.
